Use metadata to extract categories for column of specific level

For tables with data with hierarchical categories, metadata containing hierarchy level descriptions is used to extract categories at a specified level of hierarchy only.

categories_for_level(
  data,
  column_name,
  level = NA,
  strict = FALSE,
  remove_duplicates = TRUE
)

Arguments

data
data table object as returned from get_cansim()

column_name
the quoted name of the column to extract categories from

level
the hierarchy level depth to which to extract categories, where 0 is top category

strict
(default FALSE) when TRUE will only extract that specific hierarchy level

remove_duplicates
(default TRUE) When set to TRUE higher level grouping categories already captured by lower level hierarchy data will be removed

Value

A vector of categories

Examples

# \donttest{
data <- get_cansim("16-10-0117")

#> Accessing CANSIM NDM product 16-10-0117 from Statistics Canada

#> Parsing data

categories_for_level(data,"North American Industry Classification System (NAICS)",level=2)

#>  [1] "Animal food manufacturing [3111]"                                                                  
#>  [2] "Grain and oilseed milling [3112]"                                                                  
#>  [3] "Sugar and confectionery product manufacturing [3113]"                                              
#>  [4] "Fruit and vegetable preserving and specialty food manufacturing [3114]"                            
#>  [5] "Dairy product manufacturing [3115]"                                                                
#>  [6] "Meat product manufacturing [3116]"                                                                 
#>  [7] "Seafood product preparation and packaging [3117]"                                                  
#>  [8] "Bakeries and tortilla manufacturing [3118]"                                                        
#>  [9] "Other food manufacturing [3119]"                                                                   
#> [10] "Beverage manufacturing [3121]"                                                                     
#> [11] "Tobacco manufacturing [3122]"                                                                      
#> [12] "Fibre, yarn and thread mills [3131]"                                                               
#> [13] "Fabric mills [3132]"                                                                               
#> [14] "Textile and fabric finishing and fabric coating [3133]"                                            
#> [15] "Textile furnishings mills [3141]"                                                                  
#> [16] "Other textile product mills [3149]"                                                                
#> [17] "Clothing knitting mills [3151]"                                                                    
#> [18] "Cut and sew clothing manufacturing [3152]"                                                         
#> [19] "Clothing accessories and other clothing manufacturing [3159]"                                      
#> [20] "Leather and hide tanning and finishing [3161]"                                                     
#> [21] "Footwear manufacturing [3162]"                                                                     
#> [22] "Other leather and allied product manufacturing [3169]"                                             
#> [23] "Sawmills and wood preservation [3211]"                                                             
#> [24] "Veneer, plywood and engineered wood product manufacturing [3212]"                                  
#> [25] "Other wood product manufacturing [3219]"                                                           
#> [26] "Pulp, paper and paperboard mills [3221]"                                                           
#> [27] "Converted paper product manufacturing [3222]"                                                      
#> [28] "Printing and related support activities [3231]"                                                    
#> [29] "Petroleum and coal product manufacturing [3241]"                                                   
#> [30] "Basic chemical manufacturing [3251]"                                                               
#> [31] "Resin, synthetic rubber, and artificial and synthetic fibres and filaments manufacturing [3252]"   
#> [32] "Pesticide, fertilizer and other agricultural chemical manufacturing [3253]"                        
#> [33] "Pharmaceutical and medicine manufacturing [3254]"                                                  
#> [34] "Paint, coating and adhesive manufacturing [3255]"                                                  
#> [35] "Soap, cleaning compound and toilet preparation manufacturing [3256]"                               
#> [36] "Other chemical product manufacturing [3259]"                                                       
#> [37] "Plastic product manufacturing [3261]"                                                              
#> [38] "Rubber product manufacturing [3262]"                                                               
#> [39] "Clay product and refractory manufacturing [3271]"                                                  
#> [40] "Glass and glass product manufacturing [3272]"                                                      
#> [41] "Cement and concrete product manufacturing [3273]"                                                  
#> [42] "Lime and gypsum product manufacturing [3274]"                                                      
#> [43] "Other non-metallic mineral product manufacturing [3279]"                                           
#> [44] "Iron and steel mills and ferro-alloy manufacturing [3311]"                                         
#> [45] "Steel product manufacturing from purchased steel [3312]"                                           
#> [46] "Alumina and aluminum production and processing [3313]"                                             
#> [47] "Non-ferrous metal (except aluminum) production and processing [3314]"                              
#> [48] "Foundries [3315]"                                                                                  
#> [49] "Forging and stamping [3321]"                                                                       
#> [50] "Cutlery and hand tool manufacturing [3322]"                                                        
#> [51] "Architectural and structural metals manufacturing [3323]"                                          
#> [52] "Boiler, tank and shipping container manufacturing [3324]"                                          
#> [53] "Hardware manufacturing [3325]"                                                                     
#> [54] "Spring and wire product manufacturing [3326]"                                                      
#> [55] "Machine shops, turned product, and screw, nut and bolt manufacturing [3327]"                       
#> [56] "Coating, engraving, cold and heat treating and allied activities [3328]"                           
#> [57] "Other fabricated metal product manufacturing [3329]"                                               
#> [58] "Agricultural, construction and mining machinery manufacturing [3331]"                              
#> [59] "Industrial machinery manufacturing [3332]"                                                         
#> [60] "Commercial and service industry machinery manufacturing [3333]"                                    
#> [61] "Ventilation, heating, air-conditioning and commercial refrigeration equipment manufacturing [3334]"
#> [62] "Metalworking machinery manufacturing [3335]"                                                       
#> [63] "Engine, turbine and power transmission equipment manufacturing [3336]"                             
#> [64] "Other general-purpose machinery manufacturing [3339]"                                              
#> [65] "Computer and peripheral equipment manufacturing [3341]"                                            
#> [66] "Communications equipment manufacturing [3342]"                                                     
#> [67] "Audio and video equipment manufacturing [3343]"                                                    
#> [68] "Semiconductor and other electronic component manufacturing [3344]"                                 
#> [69] "Navigational, measuring, medical and control instruments manufacturing [3345]"                     
#> [70] "Manufacturing and reproducing magnetic and optical media [3346]"                                   
#> [71] "Electric lighting equipment manufacturing [3351]"                                                  
#> [72] "Household appliance manufacturing [3352]"                                                          
#> [73] "Electrical equipment manufacturing [3353]"                                                         
#> [74] "Other electrical equipment and component manufacturing [3359]"                                     
#> [75] "Motor vehicle manufacturing [3361]"                                                                
#> [76] "Motor vehicle body and trailer manufacturing [3362]"                                               
#> [77] "Motor vehicle parts manufacturing [3363]"                                                          
#> [78] "Aerospace product and parts manufacturing [3364]"                                                  
#> [79] "Railroad rolling stock manufacturing [3365]"                                                       
#> [80] "Ship and boat building [3366]"                                                                     
#> [81] "Other transportation equipment manufacturing [3369]"                                               
#> [82] "Household and institutional furniture and kitchen cabinet manufacturing [3371]"                    
#> [83] "Office furniture (including fixtures) manufacturing [3372]"                                        
#> [84] "Other furniture-related product manufacturing [3379]"                                              
#> [85] "Medical equipment and supplies manufacturing [3391]"                                               
#> [86] "Other miscellaneous manufacturing [3399]"                                                          
#> [87] "Cannabis product manufacturing [3123]"                                                             

# }