Retrieve a Statistics Canada data table using NDM catalogue number

Retrieves a data table using an NDM catalogue number as a tidy data frame. Retrieved table data is cached for the duration of the current R session only by default.

get_cansim_ndm(cansimTableNumber, language = "english",
  refresh = FALSE, timeout = 200)

Arguments

cansimTableNumber
the NDM table number to load

language
"en" or "english" for English and "fr" or "french" for French language versions (defaults to English)

refresh
(Optional) When set to TRUE, forces a reload of data table (default is FALSE)

timeout
(Optional) Timeout in seconds for downloading cansim table to work around scenarios where StatCan servers drop the network connection.

Value

tibble format data table output

Examples

get_cansim("34-10-0013")

#> Reading CANSIM NDM product 34-10-0013 from cache.

#> # A tibble: 550 x 18
#>    REF_DATE GEO   DGUID `Type of proper… UOM   UOM_ID SCALAR_FACTOR SCALAR_ID
#>    <chr>    <chr> <chr> <chr>            <chr> <chr>  <chr>         <chr>    
#>  1 2005     Cana… 2016… Residential      Doll… 81     millions      6        
#>  2 2005     Newf… 2016… Residential      Doll… 81     millions      6        
#>  3 2005     Prin… 2016… Residential      Doll… 81     millions      6        
#>  4 2005     Nova… 2016… Residential      Doll… 81     millions      6        
#>  5 2005     New … 2016… Residential      Doll… 81     millions      6        
#>  6 2005     Queb… 2016… Residential      Doll… 81     millions      6        
#>  7 2005     Onta… 2016… Residential      Doll… 81     millions      6        
#>  8 2005     Mani… 2016… Residential      Doll… 81     millions      6        
#>  9 2005     Sask… 2016… Residential      Doll… 81     millions      6        
#> 10 2005     Albe… 2016… Residential      Doll… 81     millions      6        
#> # … with 540 more rows, and 10 more variables: VECTOR <chr>, COORDINATE <chr>,
#> #   VALUE <dbl>, STATUS <chr>, SYMBOL <chr>, TERMINATED <chr>, DECIMALS <chr>,
#> #   GeoUID <chr>, `Classification Code for Type of property` <chr>, `Hierarchy
#> #   for Type of property` <chr>