Get overview list for all Statistics Canada data cubes

Generates an overview table containing metadata of available Statistics Canada data cubes.

list_cansim_cubes(lite = FALSE, refresh = FALSE)

Arguments

lite
Get the version without cube dimensions and comments for faster retrieval, default is FALSE.

refresh
Default is FALSE, repeated calls during the same session will hit the cached data. To refresh the code list during a running R session set to TRUE

Value

A tibble with available Statistics Canada data cubes, including NDM table number, cube title, start and end dates, achieve status, subject and survey codes, frequency codes and a list of cube dimensions.

Examples

# \donttest{
list_cansim_cubes()

#> Warning: The `x` argument of `as_tibble.matrix()` must have unique column names if `.name_repair` is omitted as of tibble 2.0.0.
#> Using compatibility `.name_repair`.
#> This warning is displayed once every 8 hours.
#> Call `lifecycle::last_warnings()` to see where this warning was generated.

#> # A tibble: 5,497 x 18
#>    cansim_table_nu… cubeTitleEn cubeTitleFr productId cansimId cubeStartDate
#>    <chr>            <chr>       <chr>       <chr>     <chr>    <date>       
#>  1 10-10-0001       Federal pu… Emploi du … 10100001  183-0021 1999-01-01   
#>  2 10-10-0002       Central go… Dette du g… 10100002  191-0002 2009-04-01   
#>  3 10-10-0003       Capital ma… Activités … 10100003  176-0079 1975-01-01   
#>  4 10-10-0005       Canadian C… Classifica… 10100005  385-0041 2008-01-01   
#>  5 10-10-0006       Funds adva… Avances de… 10100006  176-0091 2013-01-01   
#>  6 10-10-0007       Net new is… Émissions … 10100007  176-0082 1975-03-01   
#>  7 10-10-0008       Foreign ex… Taux de ch… 10100008  176-0067 1950-10-02   
#>  8 10-10-0009       Foreign ex… Taux de ch… 10100009  176-0064 1950-10-01   
#>  9 10-10-0010       Sales of a… Ventes de … 10100010  183-0023 2004-01-01   
#> 10 10-10-0011       Value of s… Valeur des… 10100011  183-0024 2004-01-01   
#> # … with 5,487 more rows, and 12 more variables: cubeEndDate <date>,
#> #   releaseTime <chr>, archived <lgl>, subjectCode <chr>, surveyCode <chr>,
#> #   frequencyCode <chr>, corrections <chr>, dimensions <chr>, surveyEn <chr>,
#> #   surveyFr <chr>, subjectEn <chr>, subjectFr <chr>

# }