Normalize CANSIM table values to appropriate scale

Facilitates working with CANSIM values by setting all units to counts/dollars instead of millions, etc. If "replacement_value" is not set, it will replace the VALUE field with normalized values and drop the scale column. Otherwise it will keep the scale columns and create a new column named replacement_value with the normalized value. It will attempt to parse the REF_DATE field and create an R date variable. This is currently experimental.

normalize_cansim_values(data, replacement_value = NA,
  normalize_percent = TRUE, default_month = "01", default_day = "01")

Arguments

data
A CANSIM table as returned from get_cansim() pr get_cansim_ndm()

replacement_value
(Optional) name of the column the manipulated value should be returned in. Defaults to replacing the current value field

normalize_percent
(Optional) normalize percentages by changing them to rates. TRUE by default

default_month
The default month that should be used when creating Date objects for annual data (default set to "01")

default_day
The default day of the month that should be used when creating Date objects for monthly data (default set to "01")

Value

Tibble with adjusted values