Retrieve table metadata from Statistics Canada API

Retrieves table metadata given an input table number or vector of table numbers using either the new or old table number format. Patience is suggested as the Statistics Canada API can be very slow. The `list_cansim_tables()` function can be used as an alternative to retrieve a (cached) list of CANSIM tables with (more limited) metadata.

get_cansim_cube_metadata(cansimTableNumber)

Arguments

cansimTableNumber
A new or old CANSIM/NDM table number or a vector of table numbers

Value

a tibble containing the table metadata

Examples

get_cansim_cube_metadata("34-10-0013")

#> # A tibble: 1 x 14
#>   productId cansimId cubeTitleEn cubeTitleFr cubeStartDate cubeEndDate
#>   <chr>     <chr>    <chr>       <chr>       <chr>         <chr>      
#> 1 34-10-00… 026-0018 Residentia… Valeurs de… 2005-01-01    2015-01-01 
#> # ... with 8 more variables: nbSeriesCube <chr>, nbDatapointsCube <chr>,
#> #   archiveStatusCode <chr>, archiveStatusEn <chr>, archiveStatusFr <chr>,
#> #   subjectCode <chr>, surveyCode <chr>, dimension <chr>