Retrieve Statistics Canada data table notes

Returns table notes given an NDM table number in English or French. Retrieved table information data is cached for the duration of the R session only.

get_cansim_table_notes(cansimTableNumber, language = "english",
  refresh = FALSE, timeout = 200)

Arguments

cansimTableNumber
the NDM table number to load

language
"en" or "english" for English and "fr" or "french" for French language versions (default set to English)

refresh
(Optional) When set to TRUE, forces a reload of data table (default is FALSE)

timeout
(Optional) Timeout in seconds for downloading cansim table to work around scenarios where StatCan servers drop the network connection.

Examples

get_cansim_table_notes("34-10-0013")

#> # A tibble: 3 x 2
#>   `Note ID` Note                                                                
#>   <chr>     <chr>                                                               
#> 1 1         "The methodology used in the current release differs from that used…
#> 2 2         Changes occurred in census metropolitan area geographical boundarie…
#> 3 3         Changes occurred in census metropolitan area geographical boundarie…