Search through Statistics Canada data tables

Searches through Statistics Canada data tables using a search term. A new table is generated if it already does not exist or if refresh option is set to TRUE. Search-terms are case insensitive, but will accept regular expressions for more advanced searching. The search function can search either through table titles or through table descriptions, depending on the whether or not search_description is set to TRUE or not. If refresh = TRUE, the table will be updated and regenerated using Statistics Canada's latest data. This can take some time since this process involves scraping through several hundred web pages to gather the required metadata. If option cache_path is set it will look for and store the overview table in that directory.

search_cansim_tables(search_term, search_fields = "both", refresh = FALSE)

Arguments

search_term
User-supplied search term used to find Statistics Canada data tables with matching titles

search_fields
By default, this function will search through table titles and keywords. Setting this parameter to "title" will only search through the title, setting it to "keyword" will only search through keywords

refresh
Default is FALSE, and will regenerate the table if set to TRUE

Value

A tibble with available Statistics Canada data tables, listing title, Statistics Canada data table catalogue number, deprecated CANSIM table number, description and geography that match the search term.

Examples

search_cansim_tables("Labour force")

#> Your CANSIM table overview data is 200 days old.
#> Consider setting options(cansim.cache_path="your cache path")
#> in your .Rprofile and refreshing the table via list_cansim_tables(refresh=TRUE).
#> 

#> The table won't be able to be refreshed if options(cansim.cache_path="your cache path") is not set.

#> # A tibble: 112 x 21
#>    title title_en title_fr keywords_en keywords keywords_fr notes notes_en
#>    <chr> <chr>    <chr>    <chr>       <chr>    <chr>       <chr> <chr>   
#>  1 Labo… Labour … Caracté… employment… employm… employment… Numb… Number …
#>  2 Labo… Labour … Caracté… employment… employm… employment… Numb… Number …
#>  3 Labo… Labour … Caracté… employment… employm… employment… Numb… Number …
#>  4 Labo… Labour … Caracté… employment… employm… employment… Numb… Number …
#>  5 Labo… Labour … Caracté… employment… employm… employment… Numb… Number …
#>  6 Labo… Labour … Caracté… employment… employm… employment… Numb… Number …
#>  7 Labo… Labour … Caracté… employment… employm… employment… <p>N… <p>Numb…
#>  8 Labo… Labour … Caracté… employment… employm… employment… <p>N… <p>Numb…
#>  9 Labo… Labour … Caracté… education … educati… education … <p>N… <p>Numb…
#> 10 Labo… Labour … Caracté… employment… employm… employment… Numb… Number …
#> # … with 102 more rows, and 13 more variables: notes_fr <chr>, state <chr>,
#> #   subject <chr>, date_published <chr>, frequency <chr>, revision_id <chr>,
#> #   time_period_coverage_start <chr>, time_period_coverage_end <chr>,
#> #   metadata_created <chr>, metadata_modified <chr>, url_en <chr>,
#> #   url_fr <chr>, cansim_table_number <chr>